Set up flash messages on store creation/edit
Set up the flash messages that will appear on the store page when a
store is created.

Changed the route of the rendered page after the POST request to the
store page.

Edited margin for the flash messages and textareas so that they have
enough space between elements.

Re-built css using webpack.