Create edit store page and routes
Created a template to edit a stores information by reusing the
editStore template.

Created all the controllers to handle the data.
TODO: When we add auth we need to confirm that it is the store owner
trying to edit the store information.

Created the routes for the edit store flow control.

Linted all the files so that they're uniform.